[DASH]: Miscellaneous bug fixes and adding vstests (#2745)
* [DASH]: dashrouteorch: Change routing action "route_direct" to "direct"
* [DASH]: Add vnet vstest cases
* [DASH]: dashorch: Support ENI state changes
* [DASH]: orchagent: Add PA refcount logic
* [DASH]: Add refcounting logic to support mappings where multiple CAs map to one PA
* [DASH]: Reduced logging level to avoid dropped messages
* [DASH]: Fixed setEniAdminState to add new ENI state to ENI table

+++ orchagent/dash/dashvnetorch.cpp
@@ -81,7 +81,7 @@ bool DashVnetOrch::addVnetPost(const string& vnet_name, const DashVnetBulkContex
     VnetEntry entry = { id, ctxt.guid };
     vnet_table_[vnet_name] = entry;
     gVnetNameToId[vnet_name] = id;
-    SWSS_LOG_NOTICE("Vnet entry added for %s", vnet_name.c_str());
+    SWSS_LOG_INFO("Vnet entry added for %s", vnet_name.c_str());
 
     return true;
 }
@@ -136,7 +136,7 @@ bool DashVnetOrch::removeVnetPost(const string& vnet_name, const DashVnetBulkCon
     }
     vnet_table_.erase(vnet_name);
     gVnetNameToId.erase(vnet_name);
-    SWSS_LOG_NOTICE("Vnet entry removed for %s", vnet_name.c_str());
+    SWSS_LOG_INFO("Vnet entry removed for %s", vnet_name.c_str());
 
     return true;
 }
@@ -305,12 +305,28 @@ void DashVnetOrch::addPaValidation(const string& key, VnetMapBulkContext& ctxt)
 {
     SWSS_LOG_ENTER();
 
+    auto& object_statuses = ctxt.pa_validation_object_statuses;
+    string pa_ref_key = ctxt.vnet_name + ":" + ctxt.underlay_ip.to_string();
+    auto it = pa_refcount_table_.find(pa_ref_key);
+    if (it != pa_refcount_table_.end())
+    {
+        /*
+         * PA validation entry already exisits. Just increment refcount and add
+         * a dummy success status to satisfy postop
+         */
+        object_statuses.emplace_back(SAI_STATUS_SUCCESS);
+        pa_refcount_table_[pa_ref_key]++;
+        SWSS_LOG_INFO("Increment PA refcount to %u for PA IP %s",
+                        pa_refcount_table_[pa_ref_key],
+                        ctxt.underlay_ip.to_string().c_str());
+        return;
+    }
+
     uint32_t attr_count = 1;
     sai_pa_validation_entry_t pa_validation_entry;
     pa_validation_entry.vnet_id = gVnetNameToId[ctxt.vnet_name];
     pa_validation_entry.switch_id = gSwitchId;
     swss::copy(pa_validation_entry.sip, ctxt.underlay_ip);
-    auto& object_statuses = ctxt.pa_validation_object_statuses;
     sai_attribute_t pa_validation_attr;
 
     pa_validation_attr.id = SAI_PA_VALIDATION_ENTRY_ATTR_ACTION;
@@ -319,6 +335,9 @@ void DashVnetOrch::addPaValidation(const string& key, VnetMapBulkContext& ctxt)
     object_statuses.emplace_back();
     pa_validation_bulker_.create_entry(&object_statuses.back(), &pa_validation_entry,
             attr_count, &pa_validation_attr);
+    pa_refcount_table_[pa_ref_key] = 1;
+    SWSS_LOG_INFO("Initialize PA refcount to 1 for PA IP %s",
+                    ctxt.underlay_ip.to_string().c_str());
 }
 
 bool DashVnetOrch::addVnetMap(const string& key, VnetMapBulkContext& ctxt)
@@ -340,8 +359,10 @@ bool DashVnetOrch::addVnetMap(const string& key, VnetMapBulkContext& ctxt)
         }
         return false;
     }
-    // If the VNET map is already added, don't add it to the bulker and
-    // return true so it's removed from the consumer 
+    /*
+     * If the VNET map is already added, don't add it to the bulker and
+     * return true so it's removed from the consumer
+     */
     return true;
 }
 
@@ -372,7 +393,7 @@ bool DashVnetOrch::addOutboundCaToPaPost(const string& key, const VnetMapBulkCon
             return parseHandleSaiStatusFailure(handle_status);
         }
     }
-    SWSS_LOG_NOTICE("Outbound CA to PA  map entry for %s added", key.c_str());
+    SWSS_LOG_INFO("Outbound CA to PA  map entry for %s added", key.c_str());
 
     return true;
 }
@@ -388,9 +409,12 @@ bool DashVnetOrch::addPaValidationPost(const string& key, const VnetMapBulkConte
     }
 
     auto it_status = object_statuses.begin();
+    string pa_ref_key = ctxt.vnet_name + ":" + ctxt.underlay_ip.to_string();
     sai_status_t status = *it_status++;
     if (status != SAI_STATUS_SUCCESS)
     {
+        /* PA validation entry add failed. Remove PA refcount entry */
+        pa_refcount_table_.erase(pa_ref_key);
         if (status == SAI_STATUS_ITEM_ALREADY_EXISTS)
         {
             // Retry if item exists in the bulker
@@ -404,7 +428,7 @@ bool DashVnetOrch::addPaValidationPost(const string& key, const VnetMapBulkConte
             return parseHandleSaiStatusFailure(handle_status);
         }
     }
-    SWSS_LOG_NOTICE("PA validation entry for %s added", key.c_str());
+    SWSS_LOG_INFO("PA validation entry for %s added", key.c_str());
 
     return true;
 }
@@ -416,7 +440,7 @@ bool DashVnetOrch::addVnetMapPost(const string& key, const VnetMapBulkContext& c
     bool status = addOutboundCaToPaPost(key, ctxt) && addPaValidationPost(key, ctxt);
     if (!status)
     {
-        SWSS_LOG_ERROR("addVnetMapPost return due to status false for %s ", key.c_str());
+        SWSS_LOG_ERROR("addVnetMapPost failed for %s ", key.c_str());
         return false;
     }
 
@@ -424,7 +448,7 @@ bool DashVnetOrch::addVnetMapPost(const string& key, const VnetMapBulkContext& c
     VnetMapEntry entry = {  gVnetNameToId[vnet_name], ctxt.routing_type, ctxt.dip, ctxt.underlay_ip,
         ctxt.mac_address, ctxt.metering_bucket, ctxt.use_dst_vni };
     vnet_map_table_[key] = entry;
-    SWSS_LOG_NOTICE("Vnet map added for %s", key.c_str());
+    SWSS_LOG_INFO("Vnet map added for %s", key.c_str());
 
     return true;
 }
@@ -448,13 +472,41 @@ void DashVnetOrch::removePaValidation(const string& key, VnetMapBulkContext& ctx
     SWSS_LOG_ENTER();
 
     auto& object_statuses = ctxt.pa_validation_object_statuses;
-    sai_pa_validation_entry_t pa_validation_entry;
-    pa_validation_entry.vnet_id = vnet_map_table_[key].dst_vnet_id;
-    pa_validation_entry.switch_id = gSwitchId;
-    swss::copy(pa_validation_entry.sip, vnet_map_table_[key].underlay_ip);
+    string underlay_ip = vnet_map_table_[key].underlay_ip.to_string();
+    string pa_ref_key = ctxt.vnet_name + ":" + underlay_ip;
+    auto it = pa_refcount_table_.find(pa_ref_key);
+    if (it == pa_refcount_table_.end())
+    {
+        return;
+    }
+    else
+    {
+        if (--pa_refcount_table_[pa_ref_key] > 0)
+        {
+            /*
+             * PA validation entry already exisits. Just decrement refcount and add
+             * a dummy success status to satisfy postop
+             */
+            object_statuses.emplace_back(SAI_STATUS_SUCCESS);
+            SWSS_LOG_INFO("Decrement PA refcount to %u for PA IP %s",
+                            pa_refcount_table_[pa_ref_key],
+                            underlay_ip.c_str());
+            return;
+        }
+        else
+        {
+            sai_pa_validation_entry_t pa_validation_entry;
+            pa_validation_entry.vnet_id = vnet_map_table_[key].dst_vnet_id;
+            pa_validation_entry.switch_id = gSwitchId;
+            swss::copy(pa_validation_entry.sip, vnet_map_table_[key].underlay_ip);
 
-    object_statuses.emplace_back();
-    pa_validation_bulker_.remove_entry(&object_statuses.back(), &pa_validation_entry);
+            object_statuses.emplace_back();
+            pa_validation_bulker_.remove_entry(&object_statuses.back(), &pa_validation_entry);
+            SWSS_LOG_INFO("PA refcount refcount is zero for PA IP %s, removing refcount table entry",
+                            underlay_ip.c_str());
+            pa_refcount_table_.erase(pa_ref_key);
+        }
+    }
 }
 
 bool DashVnetOrch::removeVnetMap(const string& key, VnetMapBulkContext& ctxt)
@@ -464,7 +516,7 @@ bool DashVnetOrch::removeVnetMap(const string& key, VnetMapBulkContext& ctxt)
     bool exists = (vnet_map_table_.find(key) != vnet_map_table_.end());
     if (!exists)
     {
-        SWSS_LOG_WARN("Failed to find vnet mapping %s to remove", key.c_str());
+        SWSS_LOG_INFO("Failed to find vnet mapping %s to remove", key.c_str());
         return true;
     }
 
@@ -501,7 +553,7 @@ bool DashVnetOrch::removeOutboundCaToPaPost(const string& key, const VnetMapBulk
             return parseHandleSaiStatusFailure(handle_status);
         }
     }
-    SWSS_LOG_NOTICE("Outbound CA to PA map entry for %s removed", key.c_str());
+    SWSS_LOG_INFO("Outbound CA to PA map entry for %s removed", key.c_str());
 
     return true;
 }
@@ -510,6 +562,8 @@ bool DashVnetOrch::removePaValidationPost(const string& key, const VnetMapBulkCo
 {
     SWSS_LOG_ENTER();
 
+    string underlay_ip = vnet_map_table_[key].underlay_ip.to_string();
+    string pa_ref_key = ctxt.vnet_name + ":" + underlay_ip;
     const auto& object_statuses = ctxt.pa_validation_object_statuses;
     if (object_statuses.empty())
     {
@@ -533,7 +587,8 @@ bool DashVnetOrch::removePaValidationPost(const string& key, const VnetMapBulkCo
             return parseHandleSaiStatusFailure(handle_status);
         }
     }
-    SWSS_LOG_NOTICE("PA validation entry for %s removed", key.c_str());
+
+    SWSS_LOG_INFO("PA validation entry for %s removed", key.c_str());
 
     return true;
 }
@@ -548,7 +603,7 @@ bool DashVnetOrch::removeVnetMapPost(const string& key, const VnetMapBulkContext
         return false;
     }
     vnet_map_table_.erase(key);
-    SWSS_LOG_NOTICE("Vnet map removed for %s", key.c_str());
+    SWSS_LOG_INFO("Vnet map removed for %s", key.c_str());
 
     return true;
 }
